+++ src/ups/edu/ec/controlador/Conexion.java
@@ -78,7 +78,5 @@ public class Conexion {
 		}
 
 		return lista;
-
 	}
-
 }
