+++ src/ups/edu/ec/vista/Main.java
@@ -10,18 +10,21 @@ public class Main {
 
 	public static void main(String[] args) {
 		
-		ControladorFactura cf = new ControladorFactura();
+		//ControladorFactura cf = new ControladorFactura();
 		
 		//cf.InsertarFactura("2021-01-07", (float)5.28, (float)0.72, (float)6.0, "A", "Paracetamol 20mg");
 		//cf.MostrarFacturas();
 		
-		ControladorHorario ch = new ControladorHorario();
+		//ControladorHorario ch = new ControladorHorario();
 		
-		LocalTime horaIngreso = LocalTime.parse("12:00");
-		LocalTime horaSalida = LocalTime.parse("15:00");
+		//LocalTime horaIngreso = LocalTime.parse("12:00");
+		//LocalTime horaSalida = LocalTime.parse("15:00");
 
-		ch.InsertarHorario(Time.valueOf(horaIngreso), Time.valueOf(horaSalida));
-		ch.MostrarHorario();
+		//ch.InsertarHorario(Time.valueOf(horaIngreso), Time.valueOf(horaSalida));
+		//ch.MostrarHorario();
+		
+		Index interfaces = new Index();
+		interfaces.setVisible(true);
 
 	}
